cassandra works
